no message
git-svn-id: svn://svn.code.sf.net/p/actiongame/code/trunk/ac@302 e016b1c5-ed68-44aa-a2f7-31cac88cffde

+++ source/src/bot/bot_waypoint.cpp
@@ -1428,7 +1428,7 @@ bool CWaypointClass::CanPlaceNodeHere(const vec &from)
           return false;
      }
                
-     if (fabs(s->ceil - s->floor) < player1->radius)
+     if (fabs((float)(s->ceil - s->floor)) < player1->radius)
      {
           return false;
      }
